wrap up most of the user guides
